Update 3.3
- Window Toggling broken while I work out the new structure.

- UI functionality started. AddItem(); including item stacking available for testing. 

To test, Play > select the Inventory_Player S.O. in the scripts folder > select the cog inside of the Inspector window (upper right corner) > Use Add item. 

The int exposed on Inventory_Player adds an amount to the inventory. Adjust it as you'd like. And, the definition of the included item declares how many of that item can be stacked together (x20). So, pushing 80 of the item into the inventory will create 4 stacks of 20. push 1+4+8+14 will create 2 stacks, one of 30 and one of 7.